Update index.md
Dear Luke,
Please help out so that the paragraph "hotel booking in Aarhus" either have a bullet and then the hotel name, address and phone number OR have a line of its own without the bullet. please note that the paragraph is stated twice - once in the section: "On AU funds" and once in the section: "On AUH funds". I the section: "On AUH funds" you will also find a paragraph about GoMore. Please see to that the line: If you use GoMore... gets its own line.
Thank you so much for the help with this - very much appreciated (by the way, if you find grammatical errors, please by my guest and correct those as well :-)). Best, Charlotte